Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai>=1.57.0,<2.0.0
pydantic==2.9.2
python-multipart==0.0.12
websockets==13.1
livekit-api==0.7.1
pdfplumber==0.11.4
python-docx==1.1.2